+++ src/core/component/watch/CHANGELOG.md
@@ -11,6 +11,20 @@ Changelog
 
 ## v3.0.0-rc.37 ()
 
+#### :rocket: New Feature
+
+* Added support for remote watchers
+
+```js
+this.$watch({ctx: anotherWatcher, path: foo}, () => {
+  console.log('...');
+});
+```
+
+#### :bug: Bug Fix
+
+* Fixed memory leak with watchers
+
 #### :house: Internal
 
 * Fixed ESLint warnings